+++ src/pypi_attestations/__init__.py
@@ -1,6 +1,6 @@
 """The `pypi-attestations` APIs."""
 
-__version__ = "0.0.17"
+__version__ = "0.0.18"
 
 from ._impl import (
     Attestation,